:recycle: fix : domain 폴더로 이동
