+++ src/main/java/com/kuku9/goods/domain/cart/entity/Cart.java
@@ -1,6 +1,6 @@
-package com.kuku9.goods.cart.entity;
+package com.kuku9.goods.domain.cart.entity;
 
-import com.kuku9.goods.product.entity.Product;
+import com.kuku9.goods.domain.product.entity.Product;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -9,9 +9,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import java.util.ArrayList;
-import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
