+++ src/main/java/com/kuku9/goods/domain/event/entity/Event.java
@@ -1,13 +1,13 @@
-package com.kuku9.goods.review.entity;
+package com.kuku9.goods.domain.event.entity;
 
-import com.kuku9.goods.product.entity.Product;
+import com.kuku9.goods.domain.file.entity.File;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -16,17 +16,20 @@ import lombok.NoArgsConstructor;
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
-public class Review {
+public class Event {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@Column
+	private String title;
+
+	@Column
 	private String content;
 
-	@ManyToOne
-	@JoinColumn(name = "product_id")
-	private Product product;
+	@OneToOne
+	@JoinColumn(name = "file_id")
+	private File file;
 
 }
