+++ src/main/java/com/kuku9/goods/domain/file/entity/File.java
@@ -1,32 +1,26 @@
-package com.kuku9.goods.review.comment.entity;
+package com.kuku9.goods.domain.file.entity;
 
-import com.kuku9.goods.review.entity.Review;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
-public class Comment {
+@NoArgsConstructor
+public class File {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@Column
-	private String content;
+	private String url;
 
-	@ManyToOne
-	@JoinColumn(name = "review_id")
-	private Review review;
 
 }
