+++ src/main/java/com/kuku9/goods/domain/review/comment/entity/Comment.java
@@ -1,36 +1,32 @@
-package com.kuku9.goods.event.entity;
+package com.kuku9.goods.domain.review.comment.entity;
 
-import com.kuku9.goods.file.entity.File;
+import com.kuku9.goods.domain.review.entity.Review;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Getter
 @Entity
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class Event {
+public class Comment {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@Column
-	private String title;
-
-	@Column
 	private String content;
 
-	@OneToOne
-	@JoinColumn(name = "file_id")
-	private File file;
+	@ManyToOne
+	@JoinColumn(name = "review_id")
+	private Review review;
 
 }
