+++ src/main/java/com/kuku9/goods/domain/review/entity/Review.java
@@ -1,26 +1,32 @@
-package com.kuku9.goods.file.entity;
+package com.kuku9.goods.domain.review.entity;
 
+import com.kuku9.goods.domain.product.entity.Product;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Entity
 @Getter
-@AllArgsConstructor
+@Entity
 @NoArgsConstructor
-public class File {
+@AllArgsConstructor
+public class Review {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@Column
-	private String url;
+	private String content;
 
+	@ManyToOne
+	@JoinColumn(name = "product_id")
+	private Product product;
 
 }
